requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai
openai
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
chromadb
pymysql
psycopg2-binary
